# backend/src/main/resources/api-spec.yaml
openapi: 3.0.3
info:
  title: WebElec Backend API
  version: 1.0.0
  description: API contract for the WebElec ERP backend.
servers:
  - url: http://localhost:8080
paths:
  /api/societes:
    get:
      summary: List all societies
      tags: [Societes]
      responses:
        '200':
          description: List of societies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SocieteResponse'
    post:
      summary: Create a society
      tags: [Societes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocieteRequest'
      responses:
        '200':
          description: Newly created society
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocieteResponse'
  /api/societes/{id}:
    get:
      summary: Get society by id
      tags: [Societes]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Found society
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocieteResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete society
      tags: [Societes]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/produits:
    get:
      summary: List produits
      tags: [Produits]
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProduitResponse'
    post:
      summary: Create produit
      tags: [Produits]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProduitRequest'
      responses:
        '200':
          description: Created produit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProduitResponse'
  /api/produits/{id}:
    get:
      summary: Get produit
      tags: [Produits]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Found produit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProduitResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update produit
      tags: [Produits]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProduitRequest'
      responses:
        '200':
          description: Updated produit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProduitResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete produit
      tags: [Produits]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
    SocieteRequest:
      type: object
      required: [nom, tva]
      properties:
        nom:
          type: string
          maxLength: 255
        tva:
          type: string
          maxLength: 32
        email:
          type: string
          format: email
        telephone:
          type: string
        adresse:
          type: string
          maxLength: 512
    SocieteResponse:
      allOf:
        - $ref: '#/components/schemas/SocieteRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
...
    ProduitRequest:
      type: object
      required: [reference, nom, quantiteStock, prixUnitaire, societeId]
      properties:
        reference:
          type: string
        nom:
          type: string
        description:
          type: string
        quantiteStock:
          type: integer
        prixUnitaire:
          type: number
          format: double
        societeId:
          type: integer
          format: int64
    ProduitResponse:
      allOf:
        - $ref: '#/components/schemas/ProduitRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
...